requirements.txt
# Pokemon Card Scanner — Python dependencies
# Install: pip install -r requirements.txt
opencv-python
easyocr
numpy
requests
prettytable
imagehash
Pillow
